+++ packages/router/test/lib/helpers/shared.spec.ts
@@ -1,4 +1,4 @@
-import { createRequestContext, expect, mkAddress, GAS_ESTIMATES } from "@connext/nxtp-utils";
+import { expect } from "@connext/nxtp-utils";
 import { BigNumber } from "@ethersproject/bignumber";
 import { stub } from "sinon";
 import { getNtpTimeSeconds } from "../../../src/lib/helpers";
@@ -51,50 +51,4 @@ describe("getMainnetEquivalent", () => {
   });
 });
 
-describe("calculateGasFeeInReceivingToken", () => {
-  it("should return 0 for local chains", async () => {
-    const result = await shared.calculateGasFeeInReceivingToken(
-      mkAddress("0x1"),
-      1337,
-      mkAddress("0x2"),
-      1338,
-      18,
-      createRequestContext("test"),
-    );
-    expect(result.toNumber()).to.be.eq(0);
-  });
-
-  it("should only calculate sending chain if receiving chain is not included", async () => {
-    const tokenStub = stub(shared, "getTokenPrice");
-    const gasStub = stub(shared, "getGasPrice");
-    tokenStub.onFirstCall().resolves(BigNumber.from(1));
-    tokenStub.onSecondCall().resolves(BigNumber.from(2));
-    gasStub.onFirstCall().resolves(BigNumber.from(5));
-    const result = await shared.calculateGasFeeInReceivingToken(
-      mkAddress("0x0"),
-      1,
-      mkAddress("0x2"),
-      1338,
-      18,
-      createRequestContext("test"),
-    );
-    expect(result.toNumber()).to.be.eq((5 * parseInt(GAS_ESTIMATES.fulfill) * 1) / 2);
-  });
-
-  it("should only calculate receiving chain if sending chain is not included", async () => {
-    const tokenStub = stub(shared, "getTokenPrice");
-    const gasStub = stub(shared, "getGasPrice");
-    tokenStub.onFirstCall().resolves(BigNumber.from(1));
-    tokenStub.onSecondCall().resolves(BigNumber.from(2));
-    gasStub.onFirstCall().resolves(BigNumber.from(5));
-    const result = await shared.calculateGasFeeInReceivingToken(
-      mkAddress("0x0"),
-      1338,
-      mkAddress("0x2"),
-      1,
-      18,
-      createRequestContext("test"),
-    );
-    expect(result.toNumber()).to.be.eq((5 * parseInt(GAS_ESTIMATES.prepare) * 1) / 2);
-  });
-});
+// TODO: test pass through method calculateGasFeeInReceivingToken?