+++ packages/txservice/src/chainreader.ts
@@ -1,10 +1,13 @@
-import { Signer, providers, BigNumber } from "ethers";
-import { createLoggingContext, Logger, RequestContext } from "@connext/nxtp-utils";
+import { Signer, providers, BigNumber, constants } from "ethers";
+import { createLoggingContext, GAS_ESTIMATES, Logger, RequestContext } from "@connext/nxtp-utils";
 
 import { TransactionServiceConfig, validateTransactionServiceConfig, DEFAULT_CONFIG, ChainConfig } from "./config";
 import { ReadTransaction } from "./types";
 import { ChainRpcProvider } from "./provider";
-import { ConfigurationError, ProviderNotConfigured } from "./error";
+import { ChainNotSupported, ConfigurationError, ProviderNotConfigured } from "./error";
+import { getDeployedChainIdsForGasFee, getDeployedPriceOracleContract, getPriceOracleInterface } from "./contracts";
+
+const NO_ORACLE_CHAINS: number[] = [];
 
 // TODO: I do not like that this is generally a passthrough class now - all it handles is the mapping. We should
 // probably just expose a provider getter method and have the consumer call that to access the target ChainRpcProvider
@@ -37,7 +40,7 @@ export class ChainReader {
     this.setupProviders(requestContext, signer);
   }
 
-  /// INTERFACE
+  /// CHAIN READING METHODS
   /**
    * Create a non-state changing contract call. Returns hexdata that needs to be decoded.
    *
@@ -53,15 +56,25 @@ export class ChainReader {
   }
 
   /**
-   * Gets the native asset balance for an address
+   * Gets the asset balance for a specified address for the specified chain. Optionally pass in the
+   * assetId; by default, gets the native asset.
    *
    * @param chainId - The ID of the chain for which this call is related.
    * @param address - The hexadecimal string address whose balance we are getting.
+   * @param assetId (default = ETH) - The ID (address) of the asset whose balance we are getting.
+   * @param abi - The ABI of the token contract to use for interfacing with it, if applicable (non-native).
+   * Defaults to ERC20.
+   *
    * @returns BigNumber representing the current value held by the wallet at the
    * specified address.
    */
-  public async getBalance(chainId: number, address: string): Promise<BigNumber> {
-    return await this.getProvider(chainId).getBalance(address);
+  public async getBalance(
+    chainId: number,
+    address: string,
+    assetId = constants.AddressZero,
+    abi?: string[],
+  ): Promise<BigNumber> {
+    return await this.getProvider(chainId).getBalance(address, assetId, abi);
   }
   /**
    * Get the current gas price for the chain for which this instance is servicing.
@@ -126,6 +139,229 @@ export class ChainReader {
    */
   public async getTransactionReceipt(chainId: number, hash: string): Promise<providers.TransactionReceipt> {
     return await this.getProvider(chainId).getTransactionReceipt(hash);
+  }
+
+  public async getCode(chainId: number, address: string): Promise<string> {
+    return await this.getProvider(chainId).getCode(address);
+  }
+
+  /// CONTRACT READ METHODS
+  /**
+   * Gets token price in usd from price oracle
+   *
+   * @param chainId - The network identifier.
+   * @param assetId - The asset address to get price for.
+   */
+  async getTokenPrice(chainId: number, assetId: string, _requestContext?: RequestContext): Promise<BigNumber> {
+    const { requestContext } = createLoggingContext(this.getTokenPrice.name, _requestContext);
+    const priceOracleContract = getDeployedPriceOracleContract(chainId);
+    if (!priceOracleContract) {
+      throw new ChainNotSupported(chainId.toString(), requestContext);
+    }
+    const encodedTokenPriceData = getPriceOracleInterface().encodeFunctionData("getTokenPrice", [assetId]);
+    const tokenPrice = await this.readTx({
+      chainId,
+      to: priceOracleContract.address,
+      data: encodedTokenPriceData,
+    });
+    return BigNumber.from(tokenPrice);
+  }
+
+  /**
+   * Calculates total router gas fee in token.
+   *
+   * @param sendingChainId The source chain Id
+   * @param sendingAssetId The asset address on source chain
+   * @param receivingChainId The destination chain Id
+   * @param receivingAssetId The asset address on destination chain
+   * @param _outputDecimals Decimal number of receiving asset
+   * @param requestContext Request context instance
+   */
+  async calculateGasFeeInReceivingToken(
+    sendingChainId: number,
+    sendingAssetId: string,
+    receivingChainId: number,
+    receivingAssetId: string,
+    outputDecimals: number,
+    _requestContext?: RequestContext,
+  ): Promise<BigNumber> {
+    const { requestContext, methodContext } = createLoggingContext(
+      this.calculateGasFeeInReceivingToken.name,
+      _requestContext,
+    );
+    this.logger.info("Method start", requestContext, methodContext, {
+      sendingChainId,
+      sendingAssetId,
+      receivingAssetId,
+      receivingChainId,
+      outputDecimals,
+    });
+
+    const chainIdsForGasFee = getDeployedChainIdsForGasFee();
+
+    if (!chainIdsForGasFee.includes(sendingChainId) && !chainIdsForGasFee.includes(receivingChainId)) {
+      return constants.Zero;
+    }
+    let totalCost = constants.Zero;
+    // TODO: this is returning zero when doing a rinkeby to goerli tx. i believe this is because the oracle
+    // is not configured for goerli so theres no way to translate the price to goerli
+    // TODO: we can combine these into just 2 if statements and remove the repeated logic
+    // calculate receiving token amount for gas fee
+    // if chainIdsForGasFee includes only sendingChainId, calculate gas fee for fulfill transactions
+    // if chainIdsForGasFee includes only receivingChainId, calculate gas fee for prepare transactions
+
+    const tokenPricingSendingChain = NO_ORACLE_CHAINS.includes(sendingChainId) ? 1 : sendingChainId;
+    const tokenPricingReceivingChain = NO_ORACLE_CHAINS.includes(receivingChainId) ? 1 : receivingChainId;
+
+    this.logger.info("Getting token prices", requestContext, methodContext, {
+      tokenPricingSendingChain,
+      tokenPricingReceivingChain,
+      sendingAssetId,
+      receivingAssetId,
+      outputDecimals,
+    });
+    if (chainIdsForGasFee.includes(sendingChainId)) {
+      const gasLimitForFulfill = BigNumber.from(GAS_ESTIMATES.fulfill);
+      const [ethPriceInSendingChain, receivingTokenPrice, gasPriceInSendingChain] = await Promise.all([
+        this.getTokenPrice(tokenPricingSendingChain, constants.AddressZero),
+        this.getTokenPrice(tokenPricingSendingChain, sendingAssetId),
+        this.getGasPrice(sendingChainId, requestContext),
+      ]);
+
+      // https://community.optimism.io/docs/users/fees-2.0.html#fees-in-a-nutshell
+      let l1GasInUsd = BigNumber.from(0);
+      if (sendingChainId === 10) {
+        const gasPriceMainnet = await this.getGasPrice(1, requestContext);
+        l1GasInUsd = gasPriceMainnet.mul(GAS_ESTIMATES.fulfillL1).mul(ethPriceInSendingChain);
+      }
+      const gasAmountInUsd = gasPriceInSendingChain.mul(gasLimitForFulfill).mul(ethPriceInSendingChain).add(l1GasInUsd);
+      const tokenAmountForGasFee = receivingTokenPrice.isZero()
+        ? constants.Zero
+        : gasAmountInUsd.div(receivingTokenPrice).div(BigNumber.from(10).pow(18 - outputDecimals));
+
+      totalCost = totalCost.add(tokenAmountForGasFee);
+      this.logger.info("Calculated cost on sending chain", requestContext, methodContext, {
+        totalCost: totalCost.toString(),
+        l1GasInUsd: l1GasInUsd.toString(),
+        ethPriceInSendingChain: ethPriceInSendingChain.toString(),
+        receivingTokenPrice: receivingTokenPrice.toString(),
+        gasPriceInSendingChain: gasPriceInSendingChain.toString(),
+      });
+    }
+
+    if (chainIdsForGasFee.includes(receivingChainId)) {
+      const gasLimitForPrepare = BigNumber.from(GAS_ESTIMATES.prepare);
+      const [ethPriceInReceivingChain, receivingTokenPrice, gasPriceInReceivingChain] = await Promise.all([
+        this.getTokenPrice(tokenPricingReceivingChain, constants.AddressZero),
+        this.getTokenPrice(tokenPricingReceivingChain, receivingAssetId),
+        this.getGasPrice(receivingChainId, requestContext),
+      ]);
+
+      // https://community.optimism.io/docs/users/fees-2.0.html#fees-in-a-nutshell
+      let l1GasInUsd = BigNumber.from(0);
+      if (receivingChainId === 10) {
+        const gasPriceMainnet = await this.getGasPrice(1, requestContext);
+        l1GasInUsd = gasPriceMainnet.mul(GAS_ESTIMATES.prepareL1).mul(ethPriceInReceivingChain);
+      }
+      const gasAmountInUsd = gasPriceInReceivingChain
+        .mul(gasLimitForPrepare)
+        .mul(ethPriceInReceivingChain)
+        .add(l1GasInUsd);
+      const tokenAmountForGasFee = receivingTokenPrice.isZero()
+        ? constants.Zero
+        : gasAmountInUsd.div(receivingTokenPrice).div(BigNumber.from(10).pow(18 - outputDecimals));
+
+      totalCost = totalCost.add(tokenAmountForGasFee);
+      this.logger.info("Calculated cost on receiving chain", requestContext, methodContext, {
+        totalCost: totalCost.toString(),
+        tokenAmountForGasFee: tokenAmountForGasFee.toString(),
+        l1GasInUsd: l1GasInUsd.toString(),
+        ethPriceInSendingChain: ethPriceInReceivingChain.toString(),
+        receivingTokenPrice: receivingTokenPrice.toString(),
+        gasPriceInSendingChain: gasPriceInReceivingChain.toString(),
+      });
+    }
+
+    // convert back to the intended decimals
+    return totalCost;
+  }
+
+  /**
+   * Calculates relayer fee in receiving token.
+   *
+   * @param receivingChainId The destination chain Id
+   * @param receivingAssetId The asset address on destination chain
+   * @param outputDecimals Decimal number of receiving asset
+   * @param requestContext Request context instance
+   */
+  async calculateGasFeeInReceivingTokenForFulfill(
+    receivingChainId: number,
+    receivingAssetId: string,
+    outputDecimals: number,
+    _requestContext: RequestContext,
+  ): Promise<BigNumber> {
+    const { requestContext, methodContext } = createLoggingContext(
+      this.calculateGasFeeInReceivingTokenForFulfill.name,
+      _requestContext,
+    );
+    this.logger.info("Method start", requestContext, methodContext, {
+      receivingChainId,
+      receivingAssetId,
+      outputDecimals,
+    });
+    const chaindIdsForGasFee = getDeployedChainIdsForGasFee();
+
+    if (!chaindIdsForGasFee.includes(receivingChainId)) return constants.Zero;
+    let totalCost = constants.Zero;
+
+    const tokenPricingReceivingChain = NO_ORACLE_CHAINS.includes(receivingChainId) ? 1 : receivingChainId;
+
+    if (chaindIdsForGasFee.includes(receivingChainId)) {
+      const gasLimitForFulfill = BigNumber.from(GAS_ESTIMATES.fulfill);
+      const [ethPriceInReceivingChain, receivingTokenPrice, gasPriceInReceivingChain] = await Promise.all([
+        this.getTokenPrice(tokenPricingReceivingChain, constants.AddressZero),
+        this.getTokenPrice(tokenPricingReceivingChain, receivingAssetId),
+        this.getGasPrice(receivingChainId, requestContext),
+      ]);
+
+      // https://community.optimism.io/docs/users/fees-2.0.html#fees-in-a-nutshell
+      let l1GasInUsd = BigNumber.from(0);
+      if (receivingChainId === 10) {
+        const gasPriceMainnet = await this.getGasPrice(1, requestContext);
+        l1GasInUsd = gasPriceMainnet.mul(GAS_ESTIMATES.prepareL1).mul(ethPriceInReceivingChain);
+      }
+
+      const gasAmountInUsd = gasPriceInReceivingChain
+        .mul(gasLimitForFulfill)
+        .mul(ethPriceInReceivingChain)
+        .add(l1GasInUsd);
+      const tokenAmountForGasFee = receivingTokenPrice.isZero()
+        ? constants.Zero
+        : gasAmountInUsd.div(receivingTokenPrice).div(BigNumber.from(10).pow(18 - outputDecimals));
+
+      this.logger.info("Calculated cost on receiving chain for fulfill", requestContext, methodContext, {
+        totalCost: totalCost.toString(),
+        tokenAmountForGasFee: tokenAmountForGasFee.toString(),
+        l1GasInUsd: l1GasInUsd.toString(),
+        ethPriceInReceivingChain: ethPriceInReceivingChain.toString(),
+        receivingTokenPrice: receivingTokenPrice.toString(),
+        gasPriceInReceivingChain: gasPriceInReceivingChain.toString(),
+      });
+
+      totalCost = totalCost.add(tokenAmountForGasFee);
+    }
+
+    return totalCost;
+  }
+
+  /**
+   * Helper to check for chain support gently.
+   *
+   * @param chainId - chainID of the chain to check
+   * @returns boolean indicating whether chain of chainID is supported by the service
+   */
+  public isSupportedChain(chainId: number): boolean {
+    return this.providers.has(chainId);
   }
 
   /// HELPERS